requirements.txt
Django==5.2.4
django-cors-headers==4.7.0
biopython==1.85
requests==2.31.0
pandas==2.2.2

